修改找到最小数findMin()函数的bug
size_t是unsigned类型的，传一个int类型将会导致错误

+++ chapter9/9-1/findMinNumber/findMin.cc
@@ -21,7 +21,7 @@ using namespace std;
  * @param  minNumber 输出最小值
  * @return           最小值在数组里面的位置
  */
-size_t findMin(int array[] , int arraySize , int * minNumber)
+int findMin(int array[] , int arraySize , int * minNumber)
 {
     if(array == NULL || arraySize < 0 || minNumber == NULL)
         return -1;
